services/chat-service/requirements.txt
fastapi
uvicorn
confluent-kafka
psycopg2-binary
sqlalchemy
pydantic
pydantic-settings
redis
google-generativeai
pytest
httpx
